+++ src/oihana/http/HttpParamStrategy.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace oihana\http;
+
+use oihana\reflections\traits\ConstantsTrait;
+
+/**
+ * Defines the strategy to retrieves parameters in http request (body, query or both).
+ */
+class HttpParamStrategy
+{
+    use ConstantsTrait ;
+
+    public const string BODY  = 'body' ;
+    public const string BOTH  = 'both' ;
+    public const string QUERY = 'query' ;
+}